add recursive schema module
