+++ sdks/rust/golem-rust/src/agentic/mod.rs
@@ -17,13 +17,13 @@ pub use agent_initiator::*;
 pub use agent_registry::*;
 pub use async_utils::*;
 pub use errors::*;
+pub use generic_data::*;
 pub use multimodal::*;
 pub use resolved_agent::*;
 pub use schema::*;
 pub use unstructured_binary::*;
 pub use unstructured_text::*;
 pub use wit_utils::*;
-pub use generic_type::*;
 
 mod agent;
 mod agent_impl;
@@ -31,10 +31,10 @@ mod agent_initiator;
 mod agent_registry;
 mod async_utils;
 mod errors;
+mod generic_data;
 mod multimodal;
 mod resolved_agent;
 mod schema;
 mod unstructured_binary;
 mod unstructured_text;
 mod wit_utils;
-mod generic_type;
